migrate support page to Inertia
