--- app/javascript/packs/support_portal.ts
@@ -1,9 +1,0 @@
-import ReactOnRails from "react-on-rails";
-
-import BasePage from "$app/utils/base_page";
-
-import SupportPortalPage from "$app/components/server-components/SupportPortalPage";
-
-BasePage.initialize();
-
-ReactOnRails.register({ SupportPortalPage });
