--- app/views/support/index.html.erb
@@ -1,3 +1,0 @@
-<%= load_pack("support_portal") %>
-
-<%= react_component "SupportPortalPage", props: @props, prerender: false %>
