+++ spec/controllers/support_controller_spec.rb
@@ -2,8 +2,9 @@
 
 require "spec_helper"
 require "shared_examples/authorize_called"
+require "inertia_rails/rspec"
 
-describe SupportController do
+describe SupportController, inertia: true do
   let(:seller) { create(:named_seller) }
 
 
@@ -11,17 +12,16 @@ describe SupportController do
     context "when user is signed in" do
       before { sign_in seller }
 
-      it "returns http success and assigns props with session" do
+      it "returns http success and renders Inertia component with props" do
         allow(controller).to receive(:helper_widget_host).and_return("https://help.example.test")
         allow(controller).to receive(:helper_session).and_return({ "session_id" => "abc123" })
 
         get :index
 
         expect(response).to be_successful
-        expect(assigns[:props]).to eq(
-          host: "https://help.example.test",
-          session: { "session_id" => "abc123" }
-        )
+        expect(inertia.component).to eq("Support/Index")
+        expect(inertia.props[:host]).to eq("https://help.example.test")
+        expect(inertia.props[:session]).to eq({ session_id: "abc123" })
       end
     end
 
@@ -33,6 +33,7 @@ describe SupportController do
       end
     end
   end
+
 
   describe "POST create_unauthenticated_ticket" do
     let(:valid_params) do
